export refactor to work on windows releases
