+++ .cargo/config
@@ -1,0 +1,2 @@
+[target.x86_64-pc-windows-gnu]
+linker = "x86_64-w64-mingw32-gcc"
