--- src/widgets/graph/mod.rs
@@ -1,4 +1,0 @@
-pub mod candles;
-pub mod graph;
-pub mod time_input;
-pub mod volume;
