+++ src/widgets/mod.rs
@@ -1,8 +1,11 @@
+mod candles;
 mod graph;
 mod symbols;
 mod theme;
+mod time_input;
+mod volume;
 
-pub use self::graph::graph::Graph;
-pub use self::graph::time_input::TimeInput;
+pub use self::graph::Graph;
 pub use self::symbols::Symbols;
 pub use self::theme::Theme;
+pub use self::time_input::TimeInput;
